Improve RTL support of diff view
Adds the diff-contentalign-left/right class (like in the default MediaWiki diff view), depending on whether the page (and thus the wiki) is RTL or LTR by reading the dir attribute in the mw-content-text div.

In an Ideal World, MediaWiki has a JS function to get the direction based on a language code, and we do an API request here to get the language of the page we're viewing the diff of.

I'm a javascript noob so note that there may be syntax errors even in such a simple change :)