Correct diff alignment for RTL wikis
Diffs are force-aligned to the left, so it looks weird
in RTL wikis. This patch fixes it.
If WikiLabels was loaded using ResourceLoader and CSSJanus,
this wouldn't be needed, but currently CSSJanus is not applied,
so these hacks are needed.
I tested this in the Hebrew Wikipedia and it works.